Add files via upload
I created a program called phrase_repeater where the user enters in a sentence or a phrase and the number of times the sentence or the phrase should be repeated. I decided to use a for loop to execute this program; however, a while loop can be used as well.