Add files via upload
I created a program called ball-filler, which requires the users input in order to calculate the amount of inches of cubed filler needed for each bowling ball after the user enters in how many bowling balls are being manufactured, the diameter of each ball in inches, and the core volume in inches cubed. I made sure to import the math library in the program to perform the necessary calculations.