Add files via upload
This program, called my-house, uses the graphics library to create an illustration of a house. I used the Rectangle Method for the 'house', the Polygon Method for the 'roof', and the Circle Method for the 'sun'. This program allows the user to click anywhere on the illustration to change the color of the 'sun'.