Add files via upload
I created a program called oddeven that asks a user to enter a number in order to see the number is an even number or odd number. In addition, the user is asked to enter a second number to divide the first number by, and determines if the first number divides evenly by the second number.