Add files via upload
This program was a final project of a Python course taken during my graduate studies and it works as a grade book. The user enters in the number of grades as well as each grade. The grades entered are to be formed in a list. Based on the grades that were entered, the program displays the average and the final letter grade based on the average. The program also allows the user to decide whether if he or she wishes to continue. If the user wishes to continue, he or she is allowed to make additions and modifications before displaying the average and the letter grade.